requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.0
requests==2.31.0
Pillow==10.0.0
python-dotenv==1.0.1
PyMySQL==1.1.0
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
PyJWT==2.9.0
gunicorn==21.2.0
cryptography>=42